Add `PrototypeTokenConfigPF2e`
