+++ src/module/scene/token-document/index.ts
@@ -1,2 +1,3 @@
-export { TokenConfigPF2e } from "./sheet.ts";
 export { TokenDocumentPF2e } from "./document.ts";
+export { PrototypeTokenConfigPF2e } from "./sheets/prototype-config.ts";
+export { TokenConfigPF2e } from "./sheets/token-config.ts";
